xdeepfm-attribution/requirements.txt
pyyaml
pandas
pyarrow
numpy
scikit-learn
datasets
torch>=2.0
torchmetrics>=0.11
tqdm
